+++ src/main/java/ru/task/iss/models/StatisticData.java
@@ -23,17 +23,17 @@ public class StatisticData {
     @Column(name = "count_receipts")
     private Integer countReceipts; // количество чеков
     @Column(name = "sum_without_discounts")
-    private Double sumWithoutDiscounts; //общая стоимость чеков
+    private Integer sumWithoutDiscounts; //общая стоимость чеков
     @Column(name = "avg_sum_without_discounts")
-    private Double avgSumWithoutDiscounts; //стоимость среднего чека
+    private Integer avgSumWithoutDiscounts; //стоимость среднего чека
     @Column(name = "discount_sum")
-    private Double discountSum; //сумма скидок
+    private Integer discountSum; //сумма скидок
     @Column(name = "sum_with_discounts")
-    private Double sumWithDiscount; //общая стоимость с учетом скидок
+    private Integer sumWithDiscount; //общая стоимость с учетом скидок
     @Column(name = "avg_sum_with_discounts")
-    private Double avgSumWithDiscount; //стоимость среднего чека с учетом скидок
+    private Integer avgSumWithDiscount; //стоимость среднего чека с учетом скидок
     @Column(name = "increase_receipts")
-    private Double increase; // прирост среднего чека к предыдущему часу
+    private Integer increase; // прирост среднего чека к предыдущему часу
 
     @Column(name = "newest")
     private Boolean newest;
